blowfish-mern/server/requirements.txt
pycryptodome==3.19.1
python-dotenv==1.0.1
